test: ensure createProductService return id property
